Remove old classes and update code of MLChallenge.
